feat: Move Logout button outside
