+++ src/app.css
@@ -78,3 +78,25 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+
+/* Better Scrollbars */
+
+body {
+  scrollbar-width: thin;
+  scrollbar-color: #646cff #1a1a1a;
+}
+
+body::-webkit-scrollbar {
+  width: 8px;
+  height: 8px;
+}
+
+body::-webkit-scrollbar-track {
+  background: #1a1a1a;
+}
+
+body::-webkit-scrollbar-thumb {
+  background-color: #646cff;
+  border-radius: 8px;
+  border: 3px solid #1a1a1a;
+}
