+++ src/lib/routes/Dashboard.svelte
@@ -1,30 +1,18 @@
 <script>
     // Imports
-    import { navigate } from "svelte-routing";
-    import { auth } from "../../firebase";
+    import LogoutButton from "../components/LogoutButton.svelte";
 
     export let location;
-
-    // This function logs out the user
-    const logout = async () => {
-        try {
-            await auth.signOut();
-            navigate("/");
-            console.log("User logged out");
-        } catch (error) {
-            console.error("Error while logging out:", error);
-        }
-    };
 </script>
 
 <div>
     <h1>Dashboard</h1>
     <p>Dashboard content goes here</p>
 
-    <button on:click={logout}>Log out</button>
+    <LogoutButton />
 
     <!-- Go to contexts -->
-    <a href="/context/123">Go to contexts</a>
+    <a href="/context/123">Go to an example context</a>
 </div>
 
 <style>
