+++ src/stores.js
@@ -3,8 +3,13 @@ import { auth } from "./firebase";
 
 export const loggedInUser = writable(null, function start(set) {
     const unsubscribe = auth.onAuthStateChanged((user) => {
-        set(user);
+        if (user) {
+            set(user);
+        } else {
+            set(null);
+        }
     });
+
 
     return function stop() {
         unsubscribe();
